Feat: Migration to remote ubuntu complete
SUMMARY:
--------
I reverted back to using an ordinary .env for credentials, since it was
feeling exceedingly complicated for maintanance etc (Docker swarm,
docker stack, docker secrets, etc.). I also checked the reminder
frequency is correct, to within a 3 seconds or so.

1) All credentials back in .env for both local mac and remote ubuntu
2) Migration to ubuntu complete
3) Fixed SMTP port error, needed 465, not 25 (which worked locally)
4) Fixed some permission headaches (root vs ubuntu ownership, etc, in
   particular with ragard to data/sent_events.json and an inability to
write there)
5) Contents of .env never pushed or pulled from github
6) Process was to develop locally on mace, push to github, then pull
   from remote ubuntu

+++ src/events_alerts.py
@@ -89,7 +89,7 @@ EVENT_LOOKBACK_DAYS = int(config('EVENT_LOOKBACK_DAYS', default=17))
 SCHEDULE_FREQUENCY = float(config('SCHEDULE_FREQUENCY', default=1))
 
 # Automated Reminder Frequency (days)
-REMINDER_FREQUENCY_DAYS = int(config('REMINDER_FREQUENCY_DAYS', default=30))
+REMINDER_FREQUENCY_DAYS = float(config('REMINDER_FREQUENCY_DAYS', default=30))
 
 # Timezone for scheduling & timestamps (Greece)
 LOCAL_TZ = ZoneInfo('Europe/Athens')
